Enemies move towards player constantly Added timers

+++ Code/timer.cs
@@ -1,17 +1,17 @@
 public partial class Timer {
     public bool Active { get; set; }
-    public float Amount { get; set; }
-    public float Max { get; set; }
-    public float Min { get; set; }
+    public double Amount { get; set; }
+    public double Max { get; set; }
+    public double Min { get; set; }
 
-    public Timer(float max, float min = 0, float amount = 0, bool active = true) {
+    public Timer(double max, double min = 0, double amount = 0, bool active = true) {
         Max = max;
         Min = min;
         Amount = amount;
         Active = active;
     }
 
-    public bool Process(float delta) {
+    public bool Process(double delta) {
         Amount += delta;
         return Amount >= Min && Amount > Max;
     }
